requirements.txt
requests
beautifulsoup4
schedule
PyYAML
